added exceptions list to postgres writer

+++ src/preprocessing_helper/db_misc.py
@@ -1,8 +1,12 @@
 import pandas as pd
 import psycopg2 as p
+import psycopg2.extras
 
 
-def overwrite_pg_tbl(df:pd.DataFrame, tbl_nm:str, conn:p.extensions.connection):
+def overwrite_pg_tbl(df:pd.DataFrame, tbl_nm:str, conn:p.extensions.connection, 
+                     exceptions_list):
+    if tbl_nm in exceptions_list:
+        raise Exception("Cannot write to table in exceptions_list")
     sql_columns = ",".join(df.columns)
     sql_values = "VALUES({})".format(",".join(["%s" for _ in sql_columns]))
     drop_tbl_sql = "DROP TABLE IF EXISTS {};".format(tbl_nm)
